+++ Sources/RichTextKit/Views/Pickers/RichTextAlignmentPicker.swift
@@ -13,6 +13,9 @@ import SwiftUI
 
  This view returns a plain SwiftUI `Picker` view that can be
  styled and configured in all ways supported by SwiftUI.
+
+ This picker will by default apply a segmented style, but it
+ can be disabled using `segmented: false` in the initializer.
  */
 public struct RichTextAlignmentPicker: View {
 
@@ -21,28 +24,25 @@ public struct RichTextAlignmentPicker: View {
 
      - Parameters:
        - title: The optional picker title.
-       - values: The pickable alignments, by default all available alignments.
        - selection: The binding to update with the picker.
+       - segmented: Whether or not to aply a segmented picker style, by default `true`.
+       - values: The pickable alignments, by default all available alignments.
      */
     public init(
         title: String = "",
-        values: [RichTextAlignment] = RichTextAlignment.allCases,
-        selection: Binding<RichTextAlignment>
+        selection: Binding<RichTextAlignment>,
+        segmented: Bool = true,
+        values: [RichTextAlignment] = RichTextAlignment.allCases
     ) {
         self.title = title
-        self.values = values
         self._selection = selection
+        self.segmented = segmented
+        self.values = values
     }
 
-    /**
-     The optional picker title.
-     */
-    public let title: String
-
-    /**
-     The pickable alignments.
-     */
-    public let values: [RichTextAlignment]
+    let title: String
+    let segmented: Bool
+    let values: [RichTextAlignment]
 
     @Binding
     private var selection: RichTextAlignment
@@ -52,6 +52,18 @@ public struct RichTextAlignmentPicker: View {
             ForEach(RichTextAlignment.allCases) {
                 $0.icon.tag($0)
             }
+        }.segmented(if: segmented)
+    }
+}
+
+private extension View {
+
+    @ViewBuilder
+    func segmented(if condition: Bool) -> some View {
+        if condition {
+            self.pickerStyle(.segmented)
+        } else {
+            self
         }
     }
 }